requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
coverage==7.1.0
Django==4.1.5
django-admin-honeypot-updated-2021==1.2.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-htmx==1.13.0
django-ipware==5.0.0
django-js-asset==2.0.0
django-markdown==0.8.4
django-taggit==3.1.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.6.1
gunicorn==20.1.0
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.4.1
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
zipp==3.13.0
